requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
alpha-vantage==2.3.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
billboard.py==6.2.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
COVID19Py==0.3.0
cssselect==1.1.0
fake-useragent==0.1.11
feedparser==6.0.8
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
flask-ngrok==0.0.25
Flask-WTF==0.15.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlesearch-python==1.0.1
grpcio==1.40.0
h5py==3.1.0
huggingface-hub==0.0.16
idna==2.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==0.13.2
keras==2.6.0
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.1.1
MarkupSafe==2.0.1
multidict==5.1.0
multitasking==0.0.9
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==0.25.0
parse==1.19.0
prettytable==0.7.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycricbuzz==2.4
pyee==8.2.2
pygame==2.0.1
pyparsing==2.4.7
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.21.2
scipy==1.7.1
sentencepiece==0.1.91
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
tika==1.19
tokenizers==0.10.3
torch==1.9.0
tqdm==4.62.2
transformers==4.10.2
typing-extensions==3.7.4.3
urllib3==1.26.6
w3lib==1.22.0
websockets==9.1
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
WTForms==2.3.3
yahoo-fin==0.8.9.1
yarl==1.6.3
yfinance==0.1.63
zipp==3.5.0
